requirements.txt
Django==5.0.2
gunicorn==20.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
django-filter
django-storages==1.13.2
pillow
python-dotenv
boto3==1.26.59
gunicorn==20.1.0
psycopg2-binary==2.9.6
drf-yasg
langchain==0.1.5
langchain_community==0.0.17
langchain_openai==0.0.5
dj_database_url
Jinja2==3.1.2
dj_database_url
flake8
